Move AdjointValue out of Differentiation.cpp
